+++ source/Programs/PBR/header.vert
@@ -30,8 +30,8 @@
 
 #else
 
-precision mediump float;
-precision lowp int;
+precision highp float;
+precision highp int;
 precision lowp sampler2D;
 precision lowp samplerCube;
 #if __VERSION__ >= 300
